% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.ZerosignR.result.R
\name{irf.ZerosignR.result}
\alias{irf.ZerosignR.result}
\title{Calculate structural IRFs for zero and sign restrictions estimation result}
\usage{
\method{irf}{ZerosignR.result}(zerosignr, horizon = 0, LR = FALSE, quantil = 0.05)
}
\arguments{
\item{zerosignr}{\code{ZerosignR.result} object}

\item{horizon}{Numeric. Horizon to calculate IRF for.
0 = only contemporaneous IRF, 1 = contemporaneous + next period after shock IRF, etc.}

\item{LR}{Boolean. Whether to calculate long-run IRF additionally to finite-run ones.}

\item{quantil}{Numeric. Number in (0,1), calculate IRF quantile band.}
}
\value{
\code{ZerosignR.irf}, a list containing:
\itemize{
\item \code{struc_irfs}, 3-dim. array with IRFs. The 1-st dimension is for specific Q-matrix,
the 2-nd one is \emph{nvars*(horizon+contemporaneous+LR)} --- row-stacked IRFs for different
periods, starting from the earliest one, the 3-rd one is for shocks;
\item \code{median_irf}, pointwise-median IRF, aggregated for Q-matrices;
\item \code{quantile_irf}, IRF sample quantiles for band;
\item \code{ctm_irf}, IRF from closest-to-median model, Euclidean distance;
\item \eqn{\ldots}{...}, auxillary constants.
}
}
\description{
This function calculates arbitrary-horizon structural IRFs conditinal on given
set of matrices Q --- orthonormal matrices that satisfy zero and sign restrictions.
}
\details{
Arias et al. (2018) show that structural shock IRF for period \eqn{h}{h} can be calculated in
matrix form as follows (see definitions for \eqn{J}{J} and \eqn{F}{F} in the original article):
\deqn{L_h(A_0, A_+) = \left( A_0^{-1} J^{'} F^h J \right)^{'}.}{Lh(A0, A+) = (A0 * J' * F^h * J)'.}

The special case is for long-run IRF, which is given as:
\deqn{L_\infty(A_0, A_+) = \left( A^{'}_0 - \sum_{l=1}^{p} A^{'}_l \right)^{-1}.}{Linf(A0, A+) = (A0' - ∑ A_l')^(-1).}

In order to obtain structural IRF conditional on orthonormal matrix Q, that satisfies
zero and sign restrictions, one has just multiply IRF matrix by Q:
\deqn{L_h(A_0, A_+)Q = L_h(A_0Q, A_+Q).}{Lh(A0, A+) * Q = Lh(A0 * Q, A+ * Q).}
}
\note{
\code{struc_irfs} in returned value \code{ZerosignR.irf} is a row-stacked IRF with columns for shocks.
Shocks' order is as in the restriction matrix from \code{ZerosignR.result} object.

First \emph{nvars} rows are for variables' contemporaneous response (order is as in underlying model),
Rows from \emph{nvars + 1} to \emph{2*nvars} are for variables' response in the period right after the shock, etc.
If \code{LR} is \code{TRUE}, the last \emph{nvars} rows are long-run IRF.
}
\references{
{
Arias, J.E. and Rubio-Ramirez, J. F. and Waggoner, D. F. (2018)
Inference Based on Structural Vector Autoregressions Identifiied with Sign and Zero Restrictions:
Theory and Applications. \emph{Econometrica}, 86, 2, 685-720, \url{https://doi.org/10.3982/ECTA14468}.
}
}
\seealso{
\code{\link{zerosign_restr}}
}
\author{
Artur Zmanovskii. \email{anzmanovskii@gmail.com}
}
