% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_ala_arias.R
\name{irf_ala_arias}
\alias{irf_ala_arias}
\title{Calculate structural IRFs}
\usage{
irf_ala_arias(B, Sigma, p, n, horizon, LR = FALSE, Q = NULL)
}
\arguments{
\item{B}{\eqn{B = A_+ \cdot (A_0)^(-1)}{B = A+ * A0^(-1)} --- matrix of reduced parameters in form \eqn{B = \left[c, B_1, \ldots, B_p \right]}{B = (c, B1, ..., Bp)}.
If there are no constants in model, just add extra zero column to the beginning of B.}

\item{Sigma}{Variance-covariance matrix of error term.}

\item{p}{Order of model.}

\item{n}{Number of variables/shocks, including unidentified ones.}

\item{horizon}{Number of periods to calculate IRF for. -1 => no short-run IRFs.}

\item{LR}{Boolean variable, whether to calculate long-run IRF (default=FALSE). Note if LR is TRUE, long-run IRFs.
are added to the bottom of matrix returned (the last n rows are long-run IRFs).}

\item{Q}{\emph{(optional)} Estimated orthonormal matrix used for structural IRF (default=NULL).}
}
\value{
A matrix of columnwise-stacked structural IRFs, columns = shocks (as ordered in model), rows = variables*horizon.
}
\description{
This function calculates orthogonal structural IRFs
using linear algebra primitives. Returns a \eqn{n \times (n \cdot horizon + n \cdot 1_{LR})}{n x (n*horizon + 1*LR)} matrix,
columns of which represent shocks, as ordered in model, and rows represent variables for each period of
IRF, as ordered in model, written cyclically. First n rows represent contemporaneous relations of
variables, next n rows represent same relations for h=1, and so on.
}
