% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerosign_restr_ala_arias.R
\name{zerosign_restr_ala_arias}
\alias{zerosign_restr_ala_arias}
\title{Perform simulations to identify shock under zero and sign restrictions — primitives}
\usage{
zerosign_restr_ala_arias(
  irfs,
  B,
  Sigma,
  zero_sign_matrix,
  tries,
  negative_Q = TRUE,
  perm_Q = FALSE
)
}
\arguments{
\item{irfs}{IRF matrix of size \eqn{(nvars \times (nvars*horizons))}{nvars x (nvars*horizons)} ---
columnwise-stacked IRFs by periods if there is more than one period of IRF.}

\item{B}{\eqn{B = A_+ \cdot (A_0)^(-1)}{B = A+ * A0^(-1)} --- matrix of reduced parameters in form \eqn{B = \left[c, B_1, \ldots, B_p\right]}{B = (c, B1, ..., Bp)}.
If there are no constants in model, just add extra zero column to the beginning of B.}

\item{Sigma}{Variance-covariance matrix of error term.}

\item{zero_sign_matrix}{Square matrix of zero and sign restrictions
\eqn{(nvars \times nvars)}{nvars x nvars}.
Columns = shocks, structural and residual ones. Put \emph{0} into \emph{(i,j)}-th cell of
this matrix in order to set zero restriction to the \emph{i}-th variable response to
the shock \emph{j}.}

\item{tries}{Number of tries of random orthonormal matrix generation}

\item{negative_Q}{(default = \code{TRUE}). Whether also check negative Q.}

\item{perm_Q}{NOT IMPLEMENTED. Whether to account for permutation of matrix \emph{Q}
columns to increase method's efficiency.}
}
\description{
Perform simulations to identify shock under zero and sign restrictions — primitives
}
\details{
This function simulates orthognormal random matrices \emph{Q} given number of times for
given IRF matrix in order to identify structural shocks. It returns a list of
\emph{Q} matrices and transformed IRFs that satisfy imposed restrictions.

Note that for matrix of zero and sign restrictions matrix
user should order ariables in descending order of number of zero restrictions. If there are
two or more variables with the same number of zero restrictions, their order
can be arbitrary. There cannot be more than \emph{nvars - j} zero restrictions
for \emph{j} th column (in total for all time periods).
}
